https://jira.aplana.com/browse/SBRFNDFL-2256 Рефакторинг: модуль QueryDSL генератора

+++ model/src/main/java/com/aplana/sbrf/taxaccounting/model/querydsl/QConfiguration.java
@@ -1,16 +1,15 @@
 package com.aplana.sbrf.taxaccounting.model.querydsl;
 
-import com.querydsl.core.types.Path;
+import static com.querydsl.core.types.PathMetadataFactory.*;
+
+import com.querydsl.core.types.dsl.*;
+
 import com.querydsl.core.types.PathMetadata;
-import com.querydsl.core.types.dsl.NumberPath;
-import com.querydsl.core.types.dsl.StringPath;
-import com.querydsl.sql.ColumnMetadata;
-
 import javax.annotation.Generated;
-import java.math.BigDecimal;
-import java.sql.Types;
+import com.querydsl.core.types.Path;
 
-import static com.querydsl.core.types.PathMetadataFactory.forVariable;
+import com.querydsl.sql.ColumnMetadata;
+import java.sql.Types;
 
 
 
@@ -27,7 +26,7 @@ public class QConfiguration extends com.querydsl.sql.RelationalPathBase<QConfigu
 
     public final StringPath code = createString("code");
 
-    public final NumberPath<BigDecimal> departmentId = createNumber("departmentId", BigDecimal.class);
+    public final NumberPath<Integer> departmentId = createNumber("departmentId", Integer.class);
 
     public final StringPath value = createString("value");
